avm2: Extract optimizer logic into a different file
